-Created DB model for Olejki and Oleje -Data from Olejki and Oleje is being imported into local DB

+++ app/src/main/java/com/jakubolszewski/kalkulatorwrecepturzedoz/database/Models/OlejkiModel.kt
@@ -4,5 +4,5 @@ data class OlejkiModel(
     var id: Int = 0,
     var type: String = "",
     var density: Double = 0.0,
-    var drops: Int = 0,
+    var drops: Double = 0.0,
 )
